Adding some additional stray files that are part of the metagenome linkage model.
Adding a new model of hidden linkage among reads in a metagenomic dataset.  The model and its application to metagenomic datasets is described in a manuscript in preparation by Aaron Darling and Jonathan Eisen.  This is commit phase 1: adding new files.  Phase 2 comes next, committing related changes to existing files.