Removed automatic use of 'guess trait values' during an add trait. We may not have traits embedded in taxon labels and the user may wish to load them manually or import them.
Removed the disabled name text field and type combo box from the guess traits dialog as it was unclear why these were disabled. Replaced with a label describing what is being edited.

Made the GuessTraitsDialog box persistent so the settings remain in subsequent calls (it was tedious to have to reconstruct a complex regex).