This is a very large commit. It addresses some architectural issues in BEAUti that were going to hinder implementation of continuous traits.
The main thing it does is to isolate creating and importing traits from creating partitions on those traits. A button on the data panel explicitly specifies that a trait should be used as a data partition (i.e., have a TreeLikelihood). This allows traits to be used for other purposes (such as defining Species in *BEAST).  This has also allowed considerable simplification of code and tidying up. 

There are a number of bug fixes as well over v1.6.0 that I will add to the issue list to keep track of them. Given the complex nature of these changes, I do not think these should be ported to the 1.6.0 branch but perhaps the bugs can be fixed. 

These changes will have introduced bugs and possibly cause issues with XML generation that now need to be fixed (I will attempt to catch as many as possible).