Working towards Tracer 1.6. Switched trace 'types' (in the GUI) to 'real', 'integer' and 'categorical' as the previous ones were in programmers' jargon. These buttons also are now disabled when no trace is selected. Also disabled the confirmation box for setting types - if you press the button, then you probably intended to do it (this is a static switch at the top of the TracerFrame.java).
Cleaned up some issues with the generification of 'Variate' - in particular some methods in the plotting were not being overridden.

In 'distribution' tab, for real traces, KDEs are now the default. Histogram estimates can also be shown. 

Options for plotting are now stored for specific traces (for example, if you set a manual range for a trace it is now remembered and used for this trace only).

Cleaned up some odd code in Skyline plotting (causing an erroneous warning box to appear).

Removed a few exclamation marks from dialog box messages and cleaned up the text.