adds option to not compress patterns when merging multiple patternLists
This is useful for specifying site patterns based on separate coding
regions in an alignment. The patterns can not be unique, if we want
to used robust counting methods.

The updates to Patterns do not change the API. I use arrayLists
to make sure the length of weights matches the patternLength, since this
length is used to determine the number of codons downstream. These are
casts into arrays as before for external consumption.